requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
psycopg2-binary
pymysql
pyodbc
python-dotenv
ipykernel
streamlit
plotly
